feat: add validation and blacklist classes
