+++ CHANGELOG.md
@@ -1,0 +1,8 @@
+# Changelog
+
+All notable changes to this project will be documented in this file.
+See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
+
+## Unreleased
+- Moved validation and encryption utilities to `App\Helpers\Validation`.
+- Added `App\Models\Blacklist` for IP blacklist management and removed `App\Core\Utility`.
