+++ update-api/app/Models/Blacklist.php
@@ -1,0 +1,109 @@
+<?php
+// phpcs:ignoreFile PSR1.Files.SideEffects.FoundWithSymbols
+
+/**
+ * Project: UpdateAPI
+ * Author:  Vontainment <services@vontainment.com>
+ * License: https://opensource.org/licenses/MIT MIT License
+ * Link:    https://vontainment.com
+ * Version: 3.0.0
+ *
+ * File: Blacklist.php
+ * Description: WordPress Update API
+ */
+
+namespace App\Models;
+
+class Blacklist
+{
+    /**
+     * Update the number of failed login attempts for an IP address and blacklist if necessary.
+     *
+     * @param string $ip The IP address to update.
+     * @return void
+     */
+    public static function updateFailedAttempts(string $ip): void
+    {
+        $blacklist_file = BLACKLIST_DIR . "/BLACKLIST.json";
+        $content = [];
+        if (file_exists($blacklist_file)) {
+            $raw = @file_get_contents($blacklist_file);
+            if ($raw !== false) {
+                $json = json_decode($raw, true);
+                if (is_array($json)) {
+                    $content = $json;
+                }
+            }
+        }
+
+        if (isset($content[$ip])) {
+            $content[$ip]['login_attempts'] += 1;
+            if ($content[$ip]['login_attempts'] >= 3) {
+                $content[$ip]['blacklisted'] = true;
+                $content[$ip]['timestamp'] = time();
+            }
+        } else {
+            $content[$ip] = [
+                             'login_attempts' => 1,
+                             'blacklisted'    => false,
+                             'timestamp'      => time(),
+                            ];
+        }
+
+        $fp = fopen($blacklist_file, 'c+');
+        if ($fp) {
+            if (flock($fp, LOCK_EX)) {
+                ftruncate($fp, 0);
+                rewind($fp);
+                fwrite($fp, json_encode($content));
+                fflush($fp);
+                flock($fp, LOCK_UN);
+            }
+            fclose($fp);
+        }
+    }
+
+    /**
+     * Check if an IP address is blacklisted. If the blacklist has expired, reset blacklist and login attempts.
+     *
+     * @param string $ip The IP address to check.
+     * @return bool True if the IP is blacklisted, false otherwise.
+     */
+    public static function isBlacklisted(string $ip): bool
+    {
+        $blacklist_file = BLACKLIST_DIR . "/BLACKLIST.json";
+        $blacklist = [];
+        if (file_exists($blacklist_file)) {
+            $raw = @file_get_contents($blacklist_file);
+            if ($raw !== false) {
+                $json = json_decode($raw, true);
+                if (is_array($json)) {
+                    $blacklist = $json;
+                }
+            }
+        }
+
+        if (isset($blacklist[$ip]) && $blacklist[$ip]['blacklisted']) {
+            // Check if the timestamp is older than three days
+            if (time() - $blacklist[$ip]['timestamp'] > (3 * 24 * 60 * 60)) {
+                // Remove the IP address from the blacklist and reset login_attempts
+                $blacklist[$ip]['blacklisted'] = false;
+                $blacklist[$ip]['login_attempts'] = 0;
+                $fp = fopen($blacklist_file, 'c+');
+                if ($fp) {
+                    if (flock($fp, LOCK_EX)) {
+                        ftruncate($fp, 0);
+                        rewind($fp);
+                        fwrite($fp, json_encode($blacklist));
+                        fflush($fp);
+                        flock($fp, LOCK_UN);
+                    }
+                    fclose($fp);
+                }
+            } else {
+                return true;
+            }
+        }
+        return false;
+    }
+}
